replace mimir_cluster lib with a custom module (#28)
* replace mimir_cluster lib with a custom module

* update libraries

* address Leon's comments
